Task 6 completed
